apply mock/actual separation for gpio monitor
